requirements.txt
asgiref==3.7.2
astroid==2.15.5
autopep8==2.0.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
dj-rest-auth==2.2.0
Django==4.1
django-allauth==0.52.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
Markdown==3.4.3
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.1
platformdirs==3.6.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==1.7.1
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytest==7.3.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.6.3
urllib3==2.0.3
wrapt==1.15.0
